pyproject.toml
[tool.poetry]
name = "commanderbot-lib"
version = "0.1.0-dev1"
description = "A library of utilities for discord.py bots."
authors = ["Arcensoth <arcensoth@gmail.com>"]
license = "MIT"
readme = 'README.md'
repository = "https://github.com/CommanderBot-Dev/commanderbot-lib"
homepage = "https://github.com/CommanderBot-Dev/commanderbot-lib"
keywords = ['discord', 'bot']

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
colorama = {version = "^0.4.3", optional = true}
colorlog = {version = "^4.2.1", optional = true}
pyyaml = {version = "^5.3.1", optional = true}

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.poetry.extras]
log-colors = ["colorama", "colorlog"]
yaml = ["pyyaml"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
